requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Web scraping (choose based on approach)
# selenium>=4.10.0
# playwright>=1.35.0
# beautifulsoup4>=4.12.0
# requests>=2.31.0

# API client (if using RapidAPI)
# requests>=2.31.0

# UI Framework (choose one)
# streamlit>=1.28.0
# flask>=2.3.0

# Database
# sqlalchemy>=2.0.0

# Date/Time handling
python-dateutil>=2.8.0
